Retries: Changing test name so that it is possible to run retries from Gradle Junit runner. Closes #132.
Change the testName to include the attempt. If this is not done, Gradle won't be able to run the second attempt.
It will fail since testName is used as unique identifier and there is a requirement that the same test can't run
twice. See details in https://github.com/serenity-bdd/serenity-core/issues/132